allowing the user to edit and delete only his comments

+++ app/src/main/java/com/example/foobarapplication/activities/Comment_Activity.java
@@ -24,6 +24,7 @@ public class Comment_Activity extends AppCompatActivity implements CommentListAd
     private CommentListAdapter commentListAdapter;
     private int POST_ID;
     CommentsRepository repository;
+    User user;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,7 +33,7 @@ public class Comment_Activity extends AppCompatActivity implements CommentListAd
         POST_ID = getIntent().getIntExtra("POST_ID", -1);
         repository = new CommentsRepository(this);
         Intent intentUser = getIntent();
-        User User = (com.example.foobarapplication.entities.User) intentUser.getSerializableExtra("userDetails");
+        user = (com.example.foobarapplication.entities.User) intentUser.getSerializableExtra("userDetails");
 
         RecyclerView lstComments = findViewById(R.id.lstComments);
         final EditText editTextComment = findViewById(R.id.editTextComment); // For entering new comments
@@ -53,7 +54,7 @@ public class Comment_Activity extends AppCompatActivity implements CommentListAd
             if (!commentText.isEmpty()) {
                 List<Comment> currentComments = repository.getCommentsOfPost(POST_ID); // Fetch the latest comments
                 int id = currentComments.size() + 1; // This might need adjustment based on how you manage IDs
-                Comment newComment = new Comment(id, POST_ID, User.getUserName(), commentText, User.getPhoto());
+                Comment newComment = new Comment(id, POST_ID, user.getUserName(), commentText, user.getPhoto());
                 repository.add(newComment);
                 commentListAdapter.setComments(repository.getCommentsOfPost(POST_ID)); // Refresh adapter with updated list
                 commentListAdapter.notifyItemInserted(currentComments.size() - 1);
@@ -78,46 +79,56 @@ public class Comment_Activity extends AppCompatActivity implements CommentListAd
     @Override
     public void onEditCommentClick(int position) {
         Comment commentToEdit = commentListAdapter.getCommentAt(position);
-        final EditText editTextField = new EditText(this);
-        editTextField.setText(commentToEdit.getCommentContent());
+        if (user.getUserName().equals(commentToEdit.getCommentAuthor())){
+            final EditText editTextField = new EditText(this);
+            editTextField.setText(commentToEdit.getCommentContent());
 
-        new AlertDialog.Builder(this)
-                .setTitle("Edit Comment")
-                .setView(editTextField)
-                .setPositiveButton("Update", (dialog, which) -> {
-                    String updatedCommentText = editTextField.getText().toString();
-                    if (updatedCommentText.isEmpty()) {
-                        Toast.makeText(Comment_Activity.this, "Comment cannot be empty", Toast.LENGTH_SHORT).show();
-                    } else if (updatedCommentText.equals(commentToEdit.getCommentContent())) {
-                        Toast.makeText(Comment_Activity.this, "Comment did not edit", Toast.LENGTH_SHORT).show();
-                    } else {
-                        commentToEdit.setCommentContent(updatedCommentText);
-                        commentListAdapter.notifyItemChanged(position);
-                        repository.update(commentToEdit);
-                        Toast.makeText(Comment_Activity.this, "Comment edited", Toast.LENGTH_SHORT).show();
-                    }
-                })
-                .setNegativeButton("Cancel", null)
-                .show();
+            new AlertDialog.Builder(this)
+                    .setTitle("Edit Comment")
+                    .setView(editTextField)
+                    .setPositiveButton("Update", (dialog, which) -> {
+                        String updatedCommentText = editTextField.getText().toString();
+                        if (updatedCommentText.isEmpty()) {
+                            Toast.makeText(Comment_Activity.this, "Comment cannot be empty", Toast.LENGTH_SHORT).show();
+                        } else if (updatedCommentText.equals(commentToEdit.getCommentContent())) {
+                            Toast.makeText(Comment_Activity.this, "Comment did not edit", Toast.LENGTH_SHORT).show();
+                        } else {
+                            commentToEdit.setCommentContent(updatedCommentText);
+                            commentListAdapter.notifyItemChanged(position);
+                            repository.update(commentToEdit);
+                            Toast.makeText(Comment_Activity.this, "Comment edited", Toast.LENGTH_SHORT).show();
+                        }
+                    })
+                    .setNegativeButton("Cancel", null)
+                    .show();
+        }
+        else{
+            new AlertDialog.Builder(this).setMessage("Can't edit comments of other users").show();
+        }
+
     }
 
     //delete comment button was pressed
     @Override
     public void onDeleteCommentClick(int position) {
-        new AlertDialog.Builder(this)
-                .setTitle("Delete Comment")
-                .setMessage("Are you sure you want to delete this comment?")
-                .setPositiveButton("Delete", (dialog, which) -> {
-                    // Get the comment to delete
-                    Comment commentToDelete = commentListAdapter.getCommentAt(position);
-                    // Remove from GlobalCommentsHolder
-                    repository.delete(commentToDelete);
-                    // Refresh the adapter
-                    commentListAdapter.setComments(repository.getCommentsOfPost(POST_ID));
-                    commentListAdapter.notifyItemChanged(position);
-                    Toast.makeText(Comment_Activity.this, "Comment deleted", Toast.LENGTH_SHORT).show();
-                })
-                .setNegativeButton("Cancel", null)
-                .show();
+        Comment commentToDelete = commentListAdapter.getCommentAt(position);
+        if (user.getUserName().equals(commentToDelete.getCommentAuthor())) {
+            new AlertDialog.Builder(this)
+                    .setTitle("Delete Comment")
+                    .setMessage("Are you sure you want to delete this comment?")
+                    .setPositiveButton("Delete", (dialog, which) -> {
+                        // Remove comment from room
+                        repository.delete(commentToDelete);
+                        // Refresh the adapter
+                        commentListAdapter.setComments(repository.getCommentsOfPost(POST_ID));
+                        commentListAdapter.notifyItemChanged(position);
+                        Toast.makeText(Comment_Activity.this, "Comment deleted", Toast.LENGTH_SHORT).show();
+                    })
+                    .setNegativeButton("Cancel", null)
+                    .show();
+        }
+        else{
+            new AlertDialog.Builder(this).setMessage("Can't delete comments of other users").show();
+        }
     }
 }
